Refactor runner and add CLI script to the package
